Make content pages unique on "content_book_part_id, url"
There was a uniqueness constraint on "url" and it became a problem when
a page is created from another book.  It has the same cnx url but has
different chapter sections, content book part etc.

There was talk of creating a join table like book part pages but this is
how we're going to implement this for now.