Fix task plan stats for tasked exercises that have more than one tag
For example, Chapter 14.2 of AP Biology
https://archive-staging-tutor.cnx.org/contents/e26d1433-f8e4-41db-a757-0e061d6d2737@28

We tried to look up the corresponding page for each exercise.

There are two exercises in this case, one with a lo tag and an aplo tag
("apbio-ch04-s02-lo02", "apbio-ch04-s02-aplo-2-6"), the page is only
tagged with "apbio-ch04-s02-lo02", but we were trying to search for a
page with both of those tags.

The fix is to search for a page with any of those tags, by adding
`match_count: 1` when calling Content::Routines::SearchPages.