Change research export excel file dates to strings in iso8601 format
The dates now look like this: "2016-01-22T20:03:34Z".

The failed test:

```
$ TZ=America/Chicago rspec spec/lib/export_data_spec.rb
100 / 108
F

Failures:

  1) ExportData with book exports data as a xlsx file
     Failure/Error: expect(data['First Completed At'].beginning_of_minute).to eq(

       expected: 2016-01-22 17:23:00.000000000 +0000
            got: 2016-01-22 11:23:00.000000000 +0000

       (compared using ==)

       Diff:
       @@ -1,2 +1,2 @@
       -Fri, 22 Jan 2016 17:23:00 UTC +00:00
       +2016-01-22 11:23:00 UTC

     # ./spec/lib/export_data_spec.rb:80:in `block (3 levels) in <top (required)>'

Finished in 31.97 seconds (files took 5.9 seconds to load)
1 example, 1 failure

Failed examples:

rspec ./spec/lib/export_data_spec.rb:62 # ExportData with book exports data as a xlsx file

Coverage report generated for RSpec to /home/karen/tutor-server/coverage. 3318 / 4371 LOC (75.91%) covered.
[Coveralls] Outside the CI environment, not sending data.
```

The reason for this is that the time is converted to the local / system
timezone in excel, but when it was read back out, the timezone was set to UTC.

This commit changes the date exported to a string so excel doesn't parse
it.