Add ShortCode::GetShortCodeUrl routine
If the short code points to a uri that is a task plan GID, call
Tasks::GetRedirectUrl to get the appropriate task plan url for the user.

If the short code points to a uri that is a GID but not for a task plan,
an error is returned.

Anything else is either an absolute or relative url and is returned.